% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormaker.R
\name{colormaker}
\alias{colormaker}
\title{A Colormaking function}
\usage{
colormaker(percentin)
}
\arguments{
\item{percentin}{Give a decimal value =< 1.00}
}
\description{
This function helps you go from percents to colors
}
\examples{
colormaker(.70)
}
\keyword{color}
\keyword{percent}

